packages/bottle/requirements.txt
bottle==0.12.25
argostranslate==1.8.0
python-dotenv==1.0.0